lambdas/dkms_handler/pyproject.toml
[tool.poetry]
name = "dkms-handler"
version = "0.1.0"
description = "Handle DKMS customer API requests"
authors = ["Magic Labs <team@magic.link>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.29.5"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
jwcrypto = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
